IndicatorView added to CarouselVieewDemo.xaml file
The IndicatorView control is typically placed below or above the CarouselView. The IndicatorView control displays a series of dots or other visual elements that represent the items in the CarouselView, indicating the current position within the carousel.
provides a visual representation of the items within the CarouselView and enables users to easily navigate through the carousel's content. It enhances the user experience by providing a clear indication of the current position and allowing direct access to specific items in the carousel.