Changed App icon and splash screen in GameBuddi.csproj file and their reference in the AndroidManifest.xml file.
Colors change on buttons and status bar

The .csproj file is a critical component of a .NET project that provides a central place to manage the project's configuration, dependencies, build process, and other important aspects of development and deployment.

The AndroidManifest.xml file is a crucial component of Android development that provides a comprehensive overview of the application's structure, capabilities, permissions, and interactions with the Android operating system and other applications. It plays a vital role in ensuring proper functioning, security, and integration of Android apps within the larger ecosystem.